+++ Client.Test/ItDeleteApiTest.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Configuration;
 using System.Threading.Tasks;
 using InfluxDB.Client.Api.Domain;
-using InfluxDB.Client.Core;
 using InfluxDB.Client.Writes;
 using NUnit.Framework;
 
@@ -60,16 +59,6 @@ namespace InfluxDB.Client.Test
         private DeleteApi _deleteApi;
         private Organization _organization;
         private string _token;
-
-        [Measurement("h2o")]
-        private class H20Measurement
-        {
-            [Column("location", IsTag = true)] public string Location { get; set; }
-
-            [Column("level")] public double? Level { get; set; }
-
-            [Column(IsTimestamp = true)] public DateTime Time { get; set; }
-        }
 
         [Test]
         public async Task Delete()